Move database migrations to own classes
