+++ app/src/main/java/it/niedermann/nextcloud/deck/persistence/sync/adapters/db/migration/Migration_14_15.java
@@ -1,0 +1,33 @@
+package it.niedermann.nextcloud.deck.persistence.sync.adapters.db.migration;
+
+import android.content.Context;
+
+import androidx.annotation.NonNull;
+import androidx.preference.PreferenceManager;
+import androidx.room.migration.Migration;
+import androidx.sqlite.db.SupportSQLiteDatabase;
+
+import it.niedermann.nextcloud.deck.persistence.sync.SyncWorker;
+
+public class Migration_14_15 extends Migration {
+
+    @NonNull
+    private final Context context;
+
+    public Migration_14_15(@NonNull Context context) {
+        super(14, 15);
+        this.context = context;
+    }
+
+    @Override
+    public void migrate(@NonNull SupportSQLiteDatabase database) {
+        // https://github.com/stefan-niedermann/nextcloud-deck/issues/570
+        SyncWorker.update(context);
+        // https://github.com/stefan-niedermann/nextcloud-deck/issues/525
+        PreferenceManager
+                .getDefaultSharedPreferences(context)
+                .edit()
+                .remove("it.niedermann.nextcloud.deck.theme_text")
+                .apply();
+    }
+}
